Rename PythonTestGenerationProcessor
